packages/infrastructure/pyproject.toml
[project]
name = "rag-infra"
version = "0.1.0"
description = "Infrastructure adapters for RAG pipeline"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "rag-core",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "weaviate-client>=4.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
vllm = ["vllm>=0.4.0"]
infinity = ["infinity-emb>=0.0.30"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_infra"]

[tool.uv.sources]
rag-core = { workspace = true }
